Absorb VariableAndPatternTest into PatternTest
