+++ grakn-graql/src/test/java/ai/grakn/graql/internal/query/pattern/PatternTest.java
@@ -18,11 +18,26 @@
 
 package ai.grakn.graql.internal.query.pattern;
 
+import ai.grakn.GraknGraph;
+import ai.grakn.concept.Concept;
 import ai.grakn.graql.Graql;
 import ai.grakn.graql.Pattern;
+import ai.grakn.graql.VarPattern;
+import ai.grakn.test.GraphContext;
+import ai.grakn.test.graphs.MovieGraph;
+import com.google.common.collect.Sets;
+import org.junit.ClassRule;
+import org.junit.Ignore;
+import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.ExpectedException;
+
+import java.util.Set;
+import java.util.function.Consumer;
+import java.util.stream.Collectors;
 
 import static ai.grakn.graql.Graql.and;
+import static ai.grakn.graql.Graql.neq;
 import static ai.grakn.graql.Graql.or;
 import static ai.grakn.graql.Graql.var;
 import static org.junit.Assert.assertEquals;
@@ -30,6 +45,14 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 public class PatternTest {
+
+    private GraknGraph graph = rule.graph();
+
+    @ClassRule
+    public static final GraphContext rule = GraphContext.preLoad(MovieGraph.get());
+
+    @Rule
+    public final ExpectedException exception = ExpectedException.none();
 
     @Test
     public void testVarPattern() {
@@ -43,7 +66,7 @@ public class PatternTest {
     }
 
     @Test
-    public void testDisjunction() {
+    public void testSimpleDisjunction() {
         Pattern disjunction = or();
 
         assertFalse(disjunction.admin().isVar());
@@ -55,7 +78,7 @@ public class PatternTest {
     }
 
     @Test
-    public void testConjunction() {
+    public void testSimpleConjunction() {
         Pattern conjunction = and();
 
         assertFalse(conjunction.admin().isVar());
@@ -82,5 +105,230 @@ public class PatternTest {
     public void testVarAsDisjunction() {
         //noinspection ResultOfMethodCallIgnored
         var("x").isa("movie").admin().asDisjunction();
+    }
+
+    @SuppressWarnings("ResultOfMethodCallIgnored")
+    @Test
+    public void testValidVarNames() {
+        var("123");
+        var("___");
+        var("---");
+        var("xxx");
+        var("_1");
+        var("1a");
+        var("-1");
+    }
+
+    @Ignore //TODO: FIX THIS TEST
+    @Test
+    public void whenCreatingAVarWithAnInvalidName_Throw() {
+        assertExceptionThrown(Graql::var, "");
+        assertExceptionThrown(Graql::var, " ");
+        assertExceptionThrown(Graql::var, "!!!");
+        assertExceptionThrown(Graql::var, "a b");
+        assertExceptionThrown(Graql::var, "");
+        assertExceptionThrown(Graql::var, "\"");
+        assertExceptionThrown(Graql::var, "\"\"");
+        assertExceptionThrown(Graql::var, "'");
+        assertExceptionThrown(Graql::var, "''");
+    }
+
+    @Test
+    public void testVarEquals() {
+        VarPattern var1;
+        VarPattern var2;
+
+        var1 = var().pattern();
+        var2 = var().pattern();
+        assertTrue(var1.equals(var1));
+        assertTrue(var1.equals(var2));
+
+        var1 = var("x").pattern();
+        var2 = var("y").pattern();
+        assertTrue(var1.equals(var1));
+        assertFalse(var1.equals(var2));
+
+        var1 = var("x").isa("movie");
+        var2 = var("x").isa("movie");
+        assertTrue(var1.equals(var2));
+
+        var1 = var("x").isa("movie").has("title", "abc");
+        var2 = var("x").has("title", "abc").isa("movie");
+        assertTrue(var1.equals(var2));
+    }
+
+    @Test
+    public void testConjunction() {
+        Set<VarPattern> varSet1 = Sets.newHashSet(
+                var("x").isa("movie"),
+                var("y1").isa("genre").has("name", "crime"),
+                var().isa("has-genre").rel(var("x")).rel(var("y1")));
+        Set<Concept> resultSet1 = graph.graql().match(varSet1).select("x").execute()
+                .stream()
+                .map(stringConceptMap -> stringConceptMap.get("x"))
+                .collect(Collectors.toSet());
+        assertFalse(resultSet1.isEmpty());
+
+        Set<VarPattern> varSet11 = Sets.newHashSet(var("x").pattern());
+        varSet11.addAll(varSet1);
+        Set<Concept> resultSet11 = graph.graql().match(varSet11).select("x").execute()
+                .stream()
+                .map(stringConceptMap -> stringConceptMap.get("x"))
+                .collect(Collectors.toSet());
+        assertEquals(resultSet11, resultSet1);
+
+        varSet11.add(var("z").pattern());
+        resultSet11 = graph.graql().match(varSet11).execute()
+                .stream()
+                .map(stringConceptMap -> stringConceptMap.get("x"))
+                .collect(Collectors.toSet());
+        assertEquals(resultSet11, resultSet1);
+
+        Set<VarPattern> varSet2 = Sets.newHashSet(
+                var("x").isa("movie"),
+                var("y2").isa("person").has("name", "Marlon Brando"),
+                var().isa("has-cast").rel(var("x")).rel(var("y2")));
+        Set<Concept> resultSet2 = graph.graql().match(varSet2).select("x").execute()
+                .stream()
+                .map(stringConceptMap -> stringConceptMap.get("x"))
+                .collect(Collectors.toSet());
+        assertFalse(resultSet1.isEmpty());
+
+        Set<VarPattern> varSet3 = Sets.newHashSet(
+                var("x").isa("movie"),
+                var("y").isa("genre").has("name", "crime"),
+                var().isa("has-genre").rel(var("x")).rel(var("y")),
+                var("y2").isa("person").has("name", "Marlon Brando"),
+                var().isa("has-cast").rel(var("x")).rel(var("y2")));
+        Set<Concept> conj = graph.graql().match(varSet3).select("x").execute()
+                .stream()
+                .map(stringConceptMap -> stringConceptMap.get("x"))
+                .collect(Collectors.toSet());
+        assertFalse(conj.isEmpty());
+
+        resultSet2.retainAll(resultSet1);
+        assertEquals(resultSet2, conj);
+
+        conj = graph.graql().match(and(varSet3)).execute().stream()
+                .map(stringConceptMap -> stringConceptMap.get("x"))
+                .collect(Collectors.toSet());
+        assertEquals(resultSet2, conj);
+
+        conj = graph.graql().match(or(var("x"), var("x"))).select("x").execute()
+                .stream()
+                .map(stringConceptMap -> stringConceptMap.get("x"))
+                .collect(Collectors.toSet());
+        assertTrue(conj.size() > 1);
+    }
+
+    @Test(expected = Exception.class)
+    public void whenConjunctionPassedNull_Throw() {
+        Set<VarPattern> varSet = null;
+        //noinspection ResultOfMethodCallIgnored,ConstantConditions
+        and(varSet);
+    }
+
+    @Test(expected = Exception.class)
+    public void whenConjunctionPassedVarAndNull_Throw() {
+        VarPattern var = null;
+        //noinspection ResultOfMethodCallIgnored,ConstantConditions
+        and(var(), var);
+    }
+
+    @Test
+    public void testDisjunction() {
+        Set<VarPattern> varSet1 = Sets.newHashSet(
+                var("x").isa("movie"),
+                var("y1").isa("genre").has("name", "crime"),
+                var().isa("has-genre").rel(var("x")).rel(var("y1")));
+        Set<Concept> resultSet1 = graph.graql().match(varSet1).select("x").execute()
+                .stream()
+                .map(stringConceptMap -> stringConceptMap.get("x"))
+                .collect(Collectors.toSet());
+        assertFalse(resultSet1.isEmpty());
+
+        Set<VarPattern> varSet2 = Sets.newHashSet(
+                var("x").isa("movie"),
+                var("y2").isa("person").has("name", "Marlon Brando"),
+                var().isa("has-cast").rel(var("x")).rel(var("y2")));
+        Set<Concept> resultSet2 = graph.graql().match(varSet2).select("x").execute()
+                .stream()
+                .map(stringConceptMap -> stringConceptMap.get("x"))
+                .collect(Collectors.toSet());
+        assertFalse(resultSet1.isEmpty());
+
+        Set<Pattern> varSet3 = Sets.newHashSet(
+                var("x").isa("movie"),
+                or(var("y").isa("genre").has("name", "crime"),
+                        var("y").isa("person").has("name", "Marlon Brando")),
+                var().rel(var("x")).rel(var("y")));
+        Set<Concept> conj = graph.graql().match(varSet3).select("x").execute()
+                .stream()
+                .map(stringConceptMap -> stringConceptMap.get("x"))
+                .collect(Collectors.toSet());
+        assertFalse(conj.isEmpty());
+
+        resultSet2.addAll(resultSet1);
+        assertEquals(resultSet2, conj);
+
+        conj = graph.graql().match(or(var("x"), var("x"))).select("x").execute()
+                .stream()
+                .map(stringConceptMap -> stringConceptMap.get("x"))
+                .collect(Collectors.toSet());
+        assertTrue(conj.size() > 1);
+    }
+
+    @Test(expected = Exception.class)
+    public void whenDisjunctionPassedNull_Throw() {
+        Set<VarPattern> varSet = null;
+        //noinspection ResultOfMethodCallIgnored,ConstantConditions
+        or(varSet);
+    }
+
+    @Test(expected = Exception.class)
+    public void whenDisjunctionPassedVarAndNull_Throw() {
+        VarPattern var = null;
+        //noinspection ResultOfMethodCallIgnored,ConstantConditions
+        or(var(), var);
+    }
+
+    @Test
+    public void testNegation() {
+        assertTrue(graph.graql().match(var().isa("movie").has("title", "Godfather")).ask().execute());
+        Set<Concept> result1 = graph.graql().match(
+                var("x").isa("movie").has("title", var("y")),
+                var("y").val(neq("Godfather"))).select("x").execute()
+                .stream()
+                .map(stringConceptMap -> stringConceptMap.get("x"))
+                .collect(Collectors.toSet());
+        assertFalse(result1.isEmpty());
+
+        Set<Concept> result2 = graph.graql().match(
+                var("x").isa("movie").has("title", var("y")),
+                var("y")).select("x").execute()
+                .stream()
+                .map(stringConceptMap -> stringConceptMap.get("x"))
+                .collect(Collectors.toSet());
+        assertFalse(result2.isEmpty());
+
+        result2.removeAll(result1);
+        assertEquals(1, result2.size());
+    }
+
+    @Test(expected = Exception.class)
+    public void whenNegationPassedNull_Throw() {
+        VarPattern var = null;
+        //noinspection ConstantConditions,ResultOfMethodCallIgnored
+        neq(var);
+    }
+
+    private void assertExceptionThrown(Consumer<String> consumer, String varName) {
+        boolean exceptionThrown = false;
+        try {
+            consumer.accept(varName);
+        } catch (Exception e) {
+            exceptionThrown = true;
+        }
+        assertTrue(exceptionThrown);
     }
 }
